% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att.R
\name{att}
\alias{att}
\title{Extraction of  sorted sampling/coalescent times and sampling/coalescent lineages from a phylogenetic tree.}
\usage{
att(phy, eps = 1e-06)
}
\arguments{
\item{phy}{A phylogenetic tree, which is an object of class "phylo".}

\item{eps}{Threshold parameter to separate coalescent and sampling event.}
}
\value{
A data.frame with columns t containing event times and A giving lineages.
}
\description{
\code{att} extract sorted sampling/coalescent times and sampling/coalescent lineages from a phylogenetic tree.
}
\examples{
library(ape)
t1=rcoal(20)
att(t1)


}
\references{
Palacios JA and Minin VN. Integrated nested Laplace approximation for Bayesian nonparametric phylodynamics, in Proceedings of the Twenty-Eighth Conference on Uncertainty in Artificial Intelligence, 2012.
}
\author{
Simon Frost (\email{sdwfrost@gmail.com})
}
