% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching-sampling-times.R
\name{branching.sampling.times}
\alias{branching.sampling.times}
\title{Extract sampling and coalescent times from a phylogenetic tree.}
\usage{
branching.sampling.times(phy)
}
\arguments{
\item{phy}{A phylogenetic tree.}
}
\value{
Sampling times and coalescent times
}
\description{
\code{branching.sampling.times} extracts sampling and coalescent times from a phylogenetic tree.
}
\examples{
library(ape)
t1=rcoal(20)
branching.sampling.times(t1)

}
\author{
Fei Xiang (\email{xf3087@gmail.com})
}
\references{
Palacios JA and Minin VN. Integrated nested Laplace approximation for Bayesian nonparametric phylodynamics, in Proceedings of the Twenty-Eighth Conference on Uncertainty in Artificial Intelligence, 2012.
}

