% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalgen_iso.R
\name{coalgen_iso}
\alias{coalgen_iso}
\title{Simulate coalescent times for isochronous data.}
\usage{
coalgen_iso(sample, trajectory, val_upper = 10)
}
\arguments{
\item{sample}{A two dimensional vector of number of individuals and the initial time.}

\item{trajectory}{A population growth function.}

\item{val_upper}{Upper end of time points to be simulated.}
}
\value{
Coalescent intervals and lineages.
}
\description{
\code{coalgen_iso} simulates coalescent times for isochronous data.
}
\examples{
sample<-c(100,0)

trajectory<-function(x)  exp(10*x)
example_iso<-coalgen_iso(sample, trajectory)

}
\references{
\url{https://github.com/JuliaPalacios/coalsieve}.
}
\author{
Fei Xiang (\email{xf3087@gmail.com})
}
