% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cumsum_bounded}
\alias{cumsum_bounded}
\title{Cumulative sum.}
\usage{
cumsum_bounded(x, low, high)
}
\arguments{
\item{x}{numeric vector}

\item{low}{lower bound}

\item{high}{upper bound}

\item{res}{bounded numeric vector}
}
\value{
bounded numeric vector
}
\description{
Cumulative sum.
}

