% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterochronous-gp-stat.R
\name{heterochronous.gp.stat}
\alias{heterochronous.gp.stat}
\title{Extract sampling and coalescent times from a phylogenetic tree.}
\usage{
heterochronous.gp.stat(phy)
}
\arguments{
\item{phy}{A phylogenetic tree.}
}
\value{
Sampling times and coalescent times
}
\description{
\code{heterochronous.gp.stat} extracts sampling and coalescent times from a phylogenetic tree.
}
\examples{
library(ape)
t1=rcoal(20)
heterochronous.gp.stat(t1)

}
\references{
Palacios JA and Minin VN. Integrated nested Laplace approximation for Bayesian nonparametric phylodynamics, in Proceedings of the Twenty-Eighth Conference on Uncertainty in Artificial Intelligence, 2012.
}
\author{
Fei Xiang (\email{xf3087@gmail.com})
}
