% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{loglik}
\alias{loglik}
\title{A function to calculate the log likelihood.}
\usage{
loglik(phy, parr, Model, Rcpp = F)
}
\arguments{
\item{phy}{A phylogenetic tree.}

\item{parr}{Parameters.}

\item{Model}{A Model choice from const (constant population size), expo (exponetial growth),expan (expansion growth), log (logistic growth), step (piecewise constant), pexpan (piecewise expansion growth) and plog (piecewise logistic growth).}

\item{Rcpp}{Calculation is based on C++ code when it is True and on R code when it is False.}
}
\value{
log likelihood of a phylogeny given a coalescent model and parameters.
}
\description{
A function to calculate the log likelihood.
}
\examples{
library(ape)
t1=rcoal(20)


}
\author{
Fei Xiang (\email{xf3087@gmail.com})
}

