% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{negllc_expan}
\alias{negllc_expan}
\title{Function to calculate the loglikelihood of a expansion coalescent model}
\usage{
negllc_expan(parr, t, A)
}
\arguments{
\item{parr}{A parameter vector for population size, growth rate and population size at when t is infinity.}

\item{t}{A sampling and coalescent time vector.}

\item{A}{A lineages vector}
}
\value{
loglikelihood of a expansion coalescent model.
}
\description{
Function to calculate the loglikelihood of a expansion coalescent model
}
\examples{
library(ape)
t1=rcoal(20)
x=att(t1)
negllc_expan(c(1,1,10), x$t, x$A)
library(minqa)
bobyqa(c(1,2,10),negllc_expan,lower=0,upper=Inf,t=x$t,A=x$A)
Geniefit(t1,Model="expan",start=c(10,2,2),upper=Inf,lower=0)
}
\author{
Fei Xiang (\email{xf3087@gmail.com})
}
