% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{negllc_expo}
\alias{negllc_expo}
\title{Function to calculate the loglikelihood of a exponential coalescent model}
\usage{
negllc_expo(parr, t, A)
}
\arguments{
\item{parr}{A parameter vector for population size and growth rate.}

\item{t}{A sampling and coalescent time vector.}

\item{A}{A lineages vector}
}
\value{
loglikelihood of a exponential coalescent model.
}
\description{
Function to calculate the loglikelihood of a exponential coalescent model
}
\examples{
library(ape)
t1=rcoal(20)
x=att(t1)
negllc_expo(c(1,1), x$t, x$A)
library(minqa)
bobyqa(c(1,2),negllc_expo,lower=0,upper=Inf,t=x$t,A=x$A)
Geniefit(t1,Model="expo",start=c(10,2),upper=Inf,lower=0)
}
\author{
Fei Xiang (\email{xf3087@gmail.com})
}
