% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{negllc_log}
\alias{negllc_log}
\title{Function to calculate the logistic of a logistic coalescent model}
\usage{
negllc_log(parr, t, A)
}
\arguments{
\item{parr}{A parameter vector for population size, growth rate and logistic shape parameter.}

\item{t}{A sampling and coalescent time vector.}

\item{A}{A lineages vector}
}
\value{
loglikelihood of a logistic coalescent model.
}
\description{
Function to calculate the logistic of a logistic coalescent model
}
\examples{
library(ape)
t1=rcoal(20)
x=att(t1)
negllc_log(c(1,1,10), x$t, x$A)
library(minqa)
bobyqa(c(1,2,10),negllc_log,lower=0,upper=Inf,t=x$t,A=x$A)
Geniefit(t1,Model="log",start=c(10,2,2),upper=Inf,lower=0)
library(dfoptim)
nmkb(c(1,2,10),negllc_log,lower=0,upper=Inf,t=x$t,A=x$A)
}
\author{
Fei Xiang (\email{xf3087@gmail.com})
}

