% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{negllc_plog}
\alias{negllc_plog}
\title{Function to calculate the loglikelihood of a piecewise logistic coalescent model}
\usage{
negllc_plog(parr, t, A)
}
\arguments{
\item{parr}{A parameter vector for population size, exponential growth rate and change time.}

\item{t}{A sampling and coalescent time vector.}

\item{A}{A lineages vector}
}
\value{
loglikelihood of a piecewise logistic coalescent model.
}
\description{
Function to calculate the loglikelihood of a piecewise logistic coalescent model
}
\examples{
library(ape)
t1=rcoal(20)
x=att(t1)
negllc_plog(c(1,1,.1), x$t, x$A)
library(minqa)
bobyqa(c(1,2,1),negllc_plog,lower=0,upper=Inf,t=x$t,A=x$A)
Geniefit(t1,Model="plog",start=c(10,2,1),upper=Inf,lower=0)
library(dfoptim)
nmkb(c(1,2,10),negllc_plog,lower=0,upper=Inf,t=x$t,A=x$A)
}
\author{
Fei Xiang (\email{xf3087@gmail.com})
}

