% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{negllc_step}
\alias{negllc_step}
\title{Function to calculate the loglikelihood of a step coalescent model}
\usage{
negllc_step(parr, t, A)
}
\arguments{
\item{parr}{A parameter vector for population size ,change time and propotion to population size prior to change.}

\item{t}{A sampling and coalescent time vector.}

\item{A}{A lineages vector}
}
\value{
loglikelihood of a exponential coalescent model.
}
\description{
Function to calculate the loglikelihood of a step coalescent model
}
\examples{
library(ape)
t1=rcoal(20)
x=att(t1)
negllc_step(c(1,1,.1), x$t, x$A)
library(minqa)
bobyqa(c(1,2,1),negllc_step,lower=0,upper=Inf,t=x$t,A=x$A)
Geniefit(t1,Model="step",start=c(10,2,1),upper=Inf,lower=0)
library(dfoptim)
nmkb(c(1,2,10),negllc_step,lower=0,upper=Inf,t=x$t,A=x$A)
}
\author{
Fei Xiang (\email{xf3087@gmail.com})
}

