% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att.R
\name{att}
\alias{att}
\title{Sort out sampling times, coalescent times and sampling lineages from a phylogenetic tree}
\usage{
att(phy, eps = 1e-06)
}
\description{
Sort out sampling times, coalescent times and sampling lineages from a phylogenetic tree
}

