% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox.R
\name{heterochronous.gp.stat}
\alias{heterochronous.gp.stat}
\title{Sort out sampling times, coalescent times and sampling lineages from a phylogenetic tree}
\usage{
heterochronous.gp.stat(phy)
}
\arguments{
\item{phy}{A phylogenetic tree.}
}
\value{
Sorted sampling times, coalescent times and sampling lineages.
}
\description{
\code{heterochronous.gp.stat} sorts out sampling times, coalescent times and sampling lineages from a phylogenetic tree.
}
\examples{
library(ape)
t1=rcoal(20)
heterochronous.gp.stat(t1)

}
\references{
Palacios JA and Minin VN. Integrated nested Laplace approximation for Bayesian nonparametric phylodynamics, in Proceedings of the Twenty-Eighth Conference on Uncertainty in Artificial Intelligence, 2012.
}

